+++ changes.md
@@ -12,6 +12,7 @@
    Also volume, selected for drawing in the macro, will be used in the JSROOT
 6. Support drawing of TH2Poly class with 'col' and 'lego' options
 7. Implement 'CONT', 'ARR' and 'SURF' draw options for TH2 class 
+8. Support basic drawing of TPolyLine class 
 
 
 ## Changes in 4.6.0